Actions and reducers managed in better way
