+++ app/constants/action.constant.js
@@ -1,0 +1,15 @@
+export const CATEGORY_ACTION = {
+	GET_LIST: 'GET_CATEGORY_LIST',
+	ADD: 'ADD_CATEGORY',
+	UPDATE: 'UPDATE_CATEGORY',
+	DELETE: 'DELETE_CATEGORY'
+};
+
+export const TASK_ACTION = {
+	TASK:{
+		GET_LIST: 'GET_TASK_LIST',
+		ADD: 'ADD_TASK',
+		UPDATE: 'UPDATE_TASK',
+		DELETE: 'DELETE_TASK'
+	}
+};
